feat: chat report develope(#2)
